+++ scripts/wgo-solutions/build-cp.js
@@ -11,9 +11,13 @@ const { execSync } = require("child_process");
 // };
 const build = (options) => {
   const buildPath = `./modules/${options.module}/build`;
-  fs.emptyDirSync(options.web_root);
-  console.log(`copy project build to web_root path: ${options.web_root}`);
-  fs.copySync(buildPath, options.web_root);
+  console.log(path.normalize(path.resolve(buildPath)));
+  console.log(path.normalize(options.web_root));
+  if (path.normalize(path.resolve(buildPath)) !== path.normalize(options.web_root)) {
+    fs.emptyDirSync(options.web_root);
+    console.log(`copy project build to web_root path: ${options.web_root}`);
+    fs.copySync(buildPath, options.web_root);
+  }
 };
 
 module.exports = {
